% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ids.R
\name{convert_id}
\alias{convert_id}
\title{Convert Gene Symbols to ENTREZ IDs}
\usage{
convert_id(gene_symbols)
}
\arguments{
\item{gene_symbols}{gene symbols}
}
\value{
ENTREZ gene ids
}
\description{
The function is used to convert gene symbols to entrez ids and map the genes to human orthologs
}
\examples{
convert_id(c("TP53", "A2M"))

}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
