% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_drugs.R
\name{filter_drugs}
\alias{filter_drugs}
\title{Re-rank drugs based on the number of targets}
\usage{
filter_drugs(drug_frame)
}
\arguments{
\item{drug_frame}{drugs data frame returned by `get_drugs` function}
}
\description{
The function is used to re-rank drugs based on their targets
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
