% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drugs.R
\name{get_drugs}
\alias{get_drugs}
\title{Get Drugs associated with the differential expression profile}
\usage{
get_drugs(up_regulated, down_regulated)
}
\arguments{
\item{up_regulated}{up-regulated genes returned by `prepare_ids` function}

\item{down_regulated}{down-regulated genes returned by `prepare_ids` function}
}
\description{
The function is used to get list of drugs that are associated with differential expression profile
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
