% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_drugs.R
\name{get_pert_by_type}
\alias{get_pert_by_type}
\title{Filter the drug based on the tested cell-line}
\usage{
get_pert_by_type(perts, perts_type = c("cancer", "Normal"), high_targets = TRUE)
}
\arguments{
\item{perts}{drugs data frame returned by `get_drugs` function}

\item{perts_type}{type of drug, whether it's from a normal or cancerous cell line}

\item{high_targets}{Boolean paramter to rank their drugs per the number of targets}
}
\value{
re-ranked drug data frame based on their number of targets and cell line
}
\description{
The function is used to re-rank drugs based on their targets
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
