% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_drugs.R
\name{parse_perts}
\alias{parse_perts}
\title{Parse the perturbagens to find out their frequency of occurrence.}
\usage{
parse_perts(perts)
}
\arguments{
\item{perts}{drugs data frame returned by `get_drugs` function}
}
\value{
re-ranked drug data frame based on their number of targets and cell line
}
\description{
The function is used to parse drugs based on their targets
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
