% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_drugs.R
\name{plot_activities}
\alias{plot_activities}
\title{Visualise the number of perturbagens and the top activities}
\usage{
plot_activities(perts, top = 10, directorypath = NULL)
}
\arguments{
\item{perts}{drugs data frame returned by `get_drugs` function or `get_pert_by_type` function}

\item{top}{Number of activities to be visualised (N) The default value is set to 10.}

\item{directorypath}{path to save the output figure}
}
\value{
re-ranked drug data frame based on their number of targets and cell line
}
\description{
The function is used to provide a handy visualization of the top activities
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
