% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ids.R
\name{prepare_ids}
\alias{prepare_ids}
\title{Prepare IDs for CMAP Search}
\usage{
prepare_ids(up_regulated, down_regulated)
}
\arguments{
\item{up_regulated}{up regulated gene symbols}

\item{down_regulated}{down regulated gene symbols}
}
\value{
ENTREZ gene ids
}
\description{
The function is used to prepare the ids for the CMAP search
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
