% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ids.R
\name{read_id}
\alias{read_id}
\title{Read Gene IDs from CSV file}
\usage{
read_id(csv_path)
}
\arguments{
\item{csv_path}{absolute path of CSV file containing gene symbols and sign}
}
\value{
up_regulated and down_regulated genes
}
\description{
The function is used to read IDs from a CSV file
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
